# docs/swagger.yaml
basePath: /
definitions:
  entity.Cat:
    properties:
      birthAt:
        type: string
      breed:
        type: string
      color:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  model.AddCatDto:
    properties:
      birthAt:
        type: string
      breed:
        type: string
      color:
        type: string
      name:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is a sample for Fiber Rest API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Fiber Example API
  version: "1.0"
paths:
  /api/v1/cat/{id}:
    get:
      consumes:
      - application/json
      description: Get cat by id
      parameters:
      - description: cat id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Cat'
      summary: Get cat by id
      tags:
      - cat
  /api/v1/cat/add:
    post:
      consumes:
      - application/json
      description: Add new cat data
      parameters:
      - description: Request Body
        in: body
        name: Cat
        required: true
        schema:
          $ref: '#/definitions/model.AddCatDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Cat'
      summary: Add new cat
      tags:
      - cat
  /api/v1/cat/hello:
    get:
      consumes:
      - application/json
      description: Get greeting from cat
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Say hello from cat
      tags:
      - cat
swagger: "2.0"